+++ firebase-vscode/src/test/integration/fishfood/emulator.ts
@@ -1,6 +1,9 @@
 import { firebaseSuite, firebaseTest } from "../../utils/test_hooks";
 import { FirebaseCommands } from "../../utils/page_objects/commands";
 import { FirebaseSidebar } from "../../utils/page_objects/sidebar";
+
+import { TerminalView } from "../../utils/page_objects/terminal";
+import { Notifications } from "../../utils/page_objects/notifications";
 import { mockUser } from "../../utils/user";
 import { mockProject } from "../../utils/projects";
 
@@ -26,6 +29,59 @@ firebaseSuite("Emulators", async function () {
       const current = await sidebar.currentEmulators();
 
       expect(current).toContain("dataconnect :9399");
+    },
+  );
+
+  firebaseTest(
+    "Clicking on `Clear Data Connect data` triggers a clear data event",
+    async function () {
+      const workbench = await browser.getWorkbench();
+
+      const sidebar = new FirebaseSidebar(workbench);
+      const commands = new FirebaseCommands();
+      const terminal = new TerminalView(workbench);
+
+      await sidebar.openExtensionSidebar();
+      await commands.waitForUser();
+
+      await mockUser({ email: "test@gmail.com" });
+      await mockProject("test-project");
+
+      await sidebar.startEmulators();
+      console.log("Waiting for emulators to start...");
+      await commands.waitForEmulators();
+
+      await sidebar.clearEmulatorData();
+      const text = await terminal.getTerminalText();
+      expect(text.includes("Clearing data from Data Connect data sources")).toBeTruthy();
+      
+    },
+  );
+
+  firebaseTest(
+    "Clicking on `Export emulator data` triggers a export data event",
+    async function () {
+      const workbench = await browser.getWorkbench();
+
+      const sidebar = new FirebaseSidebar(workbench);
+      const commands = new FirebaseCommands();
+      const notifications = new Notifications(workbench);
+
+      await sidebar.openExtensionSidebar();
+      await commands.waitForUser();
+
+      await mockUser({ email: "test@gmail.com" });
+      await mockProject("test-project");
+
+      await sidebar.startEmulators();
+      console.log("Waiting for emulators to start...");
+      await commands.waitForEmulators();
+
+      await sidebar.exportEmulatorData();
+
+      // runEmulatorsExport
+      const exportNotification = await notifications.getExportNotification();
+      expect(exportNotification).toExist();
     },
   );
 });