+++ firebase-vscode/CHANGELOG.md
@@ -1,5 +1,7 @@
 ## NEXT
 
+- [Added] Added support for emulator import/export.
+
 ## 0.12.0
 
 - Updated internal firebase-tools dependency to 13.29.1